Split writer-specific tests out of test_qrcode
